+++ src/App.tsx
@@ -1,7 +1,6 @@
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { ReactQueryProvider, FiltersProvider } from "providers";
 import { queryClient } from "lib";
-import { invoicesLoader } from "utils";
 import { LayoutInvoices, LayoutInvoice, LayoutShared } from "layouts";
 import {
   ErrorPage,
@@ -10,6 +9,7 @@ import {
   InvoiceEdit,
   InvoiceCreate,
   InvoiceView,
+  invoicesLoader,
 } from "pages";
 
 import "./App.css";
