+++ src/utils/index.ts
@@ -2,4 +2,3 @@ export { capitalize } from "./helpers";
 export { paginateInvoices } from "./helpers";
 export { filterInvoices } from "./helpers";
 export { getTotalPages } from "./helpers";
-export { invoicesLoader } from "./loaders";
